Add Quotes controller #edit and #new
Making use fo the same partial.

sanitize("&larr; Back to quotes"): This part of the code is using the
sanitize helper method, which is used to escape potentially dangerous
HTML tags and characters to prevent cross-site scripting (XSS) attacks.
In this case, it's sanitizing the text "← Back to quotes" before it's
displayed in the hyperlink.