Set up simple_form
Simple form helps us to define wrappers for forms which we can easily
reuse across the application (and target the same css classes). It also
helps to set form labels, defined in a separate yml file.

The autofocus is to focus the input field when the form appears on the
screen, do we don’t have to use the mouse to type directly into it.

*setting up simple_form*
1. add to Gemfile
2. bundle
3. rails g simple_form:install

Here, we've added a custom config specified in the tutorial. In the yml
file we're able to define reusable form labels.