Add :status_unprocessable_entity to #create and #edit
This is a breaking change since Rails 7 due to Turbo Drive - need to add
the line status: unprocessable_entity.

Setting the status to :unprocessable_entity in the response ensures that
the client-side Turbo JavaScript knows that the form submission was
unsuccessful due to validation errors. This allows the JavaScript to
handle the response accordingly, typically by displaying the validation
errors to the user without a full page reload.