Use Cuprite gem for system tests
Prevents chrome opening in the browser and prevents the following error in
CI:

Selenium::WebDriver::Error::SessionNotCreatedError:
            session not created: Chrome failed to start: exited normally.
              (session not created: DevToolsActivePort file doesn't exist)
              (The process started from chrome location
               /usr/bin/google-chrome is no longer running, so ChromeDriver
               is assuming that Chrome has crashed.)

The Cuprite gem in Rails is a headless driver for Capybara powered by
PhantomJS. It provides an alternative to using Selenium WebDriver with
Chrome or Firefox, allowing you to run feature/system tests in a headless
manner. Cuprite uses the PhantomJS browser internally, which means it
doesn't require an actual graphical browser window to be opened during
test execution.

Uses CB setup in spec/support file.