Add the media mixin
What is a mixin?

Mixins allow you to avoid repetitive code by encapsulating common styles
and applying them to multiple elements throughout your stylesheets.

@media (min-width: 50rem) { @content; }:

If this line generates a media query for screen sizes with a min width
of 50rem (tablets and larger). @content is a placeholder that allows the
mixin to include the block of styles passed to it.

**Example use••

scss component

.container {
  background-color: lightgray;
  color: black;

  @include media(tabletAndUp) {
    background-color: darkgray;
    color: white;
  }
}

generated css

.container {
  background-color: lightgray;
  color: black;
}

@media (min-width: 50rem) {
  .container {
    background-color: darkgray;
    color: white;
  }
}

It allows you to write more maintainable and readable media queries by
encapsulating the media query logic within a reusable mixin.

**Why start with mobile?**

mobile first approach - start with the smallest screen, then add overrides
for larger screens.