Add config folder with variables file
This is where we will choose colours, fonts and a spaces scale to ensure
consistent spacing.

Can use the multiple of 8 rule - use values that are multiples of 8 for
margins, padding, font sizes, and other spacing-related properties. You
can also use a design 'personality' which guides the fonts and colours.

The first set of rules added here ensures consistent text design.