[Experiment] You can target the _top frame
Every page has the _top frame - it's the whole page. If we set the
data-turbo-frame to _top, it replaces the whole source page with the
whole target page (looking like what we had before.). It also changes
the url to the target page which is different to a normal Turbo Frame.