Use the _top for the delete method
The same applies from the previous commit, to the delete method. Remember
that because this is a form submission, the data-turbo-frame attribute
must sit within the form data. When the delete action is executed, it
replaces the whole page with a redirect back to to the Quote#index.

Adding frames for the #show, #edit and #delete action has had unintended
consequences in that if you open the #edit form and then delete the quote
below, the #edit form is closed. A problem that can be solved with
Turbo Streams...