Implement the Turbo Stream format for #delete
Forms use this as standard in Rails 7:

Started DELETE "/quotes/908005781" for 127.0.0.1 at 2022-01-27 15:30:13 +0100
Processing by QuotesController#destroy as TURBO_STREAM

We can use this format to only make our #destroy action remove the frame
and not cause a full page replacement.

In the controller we need to support both HTML and Turbo Stream. As with
any other format the `format.turbo_stream` needs a corresponding view -
IMPORTANT.

If we take a look at this view in the network tab of the request
under '908005758':

<turbo-stream action="remove" target="quote_908005758"></turbo-stream>

As we can see, the turbo_stream helper creates the custom turbo-stream
element in the HTML. It receives the method #remove which removes the
Turbo Frame corresponding to the deleted quote - leaving the rest of
the page untouched.