[New quote] Wrap #new form in frame and target it in 'New Quote' link
Here, we:
1. remove the old frame wrapped around the header
2. add an empty `new_quote` frame (dom_id(Quote.new) is the same as
  `new_quote`)
2. wrap the form on the #new page in a frame with the same ID ad the
   empty frame in the #index page

With everything working, clicking the "New quote" button will replace the
empty content of the `new_quote` frame in the #index with the form to
create a new quote. More detailed explanation of this in the tutorial.

In the #new, @quote is a new record so the following are equivalent:
- turbo_frame_tag "new_quote"
- turbo_frame_tag Quote.new
- turbo_frame_tag @quote

Remember that `turbo_frame_tag` already applies the dom_id method so we
don't need to explicitly add it.