Prepend the new quote to the list of quotes
Currently, when clicking the 'Create quote' button with a valid input,
the quote is being created in the db but not being added to the page.To do
this we need to create a Turbo Stream view and change the behaviour in
the Quotes#create action.

What is going on in the view?
- the first line tells Turbo to prepend the to the frame with ID "quotes"
  the quotes/quote partial with the newly created @quote.
- the second line tells turbo to update the empty frame (which was being
  used to house the Quotes#new form) with empty content so that it disappears.

This is what it looks like in if we look in the network tab in the browser:

<turbo-stream action="prepend" target="quotes">
  <template>
    <turbo-frame id="quote_123">
      <!-- The HTML for the quote partial -->
    <turbo-frame>
  </template>
</turbo-stream>

<turbo-stream action="update" target="new_quote">
  <template>
    <!-- An empty template! -->
  </template>
</turbo-stream>

NB. There's another syntax that could've been used for prepending which
is good for when the content is too long:

<%= turbo_stream.prepend "quotes" do %>
  <%= render partial: "quotes/quote", locals: { quote: @quote } %>
<% end %>