[Refactor] Refactor using, RoR partial rendering shorthand
<%= turbo_stream.prepend "quotes" do %>
  <%= render partial: "quotes/quote", locals: { quote: @quote } %>
<% end %>

becomes...

<%= turbo_stream.prepend "quotes" do %>
  <%= render @quote %>
<% end %>

becomes...

<%= turbo_stream.prepend "quotes", @quote %>

Why?

This is a shorthand provided by Rails for rendering a single object.
When you pass an object to render, Rails performs the following steps:

It infers the partial name from the object's class. In this case,
if @quote is an instance of the Quote model, Rails looks for a partial
named _quote.html.erb in the app/views/quotes/ directory.

It automatically assigns the object to a local variable with the same
name as the object's class, but in lowercase. In this case, it creates
a local variable quote for the partial.