Add fixtures for User
In this example, an eavesdropper at Cos is listening in on coffee
secrets discussed between an barista and their manager. What use are these
secrets to the Cos employee? Well it turns out the Cos employee belongs to
one of the three founding families of Brighton, this one owning 75% of the
'independent' coffee shops across Brighton & Hove. They have been placed to
learn the secrets of the coffee giants, starting with Pret.