Fix CI issue - assets:precompile defaults to development db
For some reason, since adding Users (via Devise), the `assets:precompile`
task CI fails because it defaults to a development database which is not
configured:

Run bundle exec rake assets
rake aborted!
ActiveRecord::AdapterNotSpecified: The development database is not configured for the development environment.

This seems to have been fixed by manually specifying the
RAILS_ENV var at this step in the CI script.